+++ src/abc/Renderer.js
@@ -40,7 +40,7 @@ export class Renderer extends Component {
 				data-value={value} // Passing value (to eventually work with in the external wrapper)
 				data-key={key} // Passing key (to eventually work with in the external wrapper)
 				component={params?.WrapperComponent || "span"}
-				>
+			>
 				{value}
 			</RendererWrapper>
 		);
