+++ src/utils/textFormatting/highlightChildren.jsx
