+++ src/utils/textFormatting/highlightSearchedFulltexts.js
