+++ src/utils/textFormatting/highlightSearchedFulltexts.scss
